# services/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zenith API
  description: |
    Zenith is an open-core Platform-as-a-Service (PaaS) for deploying apps,
    databases, and storage on Kubernetes. This spec covers public, protected,
    admin, webhook, internal, and legacy endpoints.
  version: 1.0.0
  contact:
    name: DoTech
    url: https://freezenith.com
  license:
    name: BSL 1.1
    url: https://github.com/dotechhq/zenith/blob/main/LICENSE
servers:
  - url: https://api.freezenith.com
    description: Production
  - url: http://127.0.0.1:8080
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /api/v1/auth/login
    InternalSecret:
      type: apiKey
      in: header
      name: X-Internal-Secret
      description: Shared secret for internal metering endpoints
    WebhookSignature:
      type: apiKey
      in: header
      name: X-Hub-Signature-256
      description: HMAC-SHA256 signature for inbound webhooks
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
    Forbidden:
      description: Forbidden — insufficient role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
...
  schemas:
...
    APIError:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: invalid request body
        detail:
          type: string
      required: [code, message]
    Message:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
      required: [refresh_token]
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          description: Seconds until access token expires
          example: 3600
...
    MFACodeRequest:
      type: object
      properties:
        code:
          type: string
      required: [code]
    MFABackupCodesResponse:
      type: object
      properties:
        backup_codes:
          type: array
          items:
            type: string

    # ── User Webhooks ──
...
    Permission:
      type: string
      enum: [deploy, view_logs, manage_db, manage_team, manage_billing, admin]
    CreateRoleInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
      required: [name, permissions]
    CustomRole:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        name:
          type: string
        description:
          type: string
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AssignRoleInput:
      type: object
      properties:
        member_id:
          type: string
      required: [member_id]
    RoleAssignment:
      type: object
      properties:
        id:
          type: string
        role_id:
          type: string
        member_id:
          type: string
        assigned_by:
          type: string
        created_at:
          type: string
          format: date-time

    # ── IP Whitelist ──
...
paths:
...
  /api/v1/auth/refresh:
    post:
      operationId: refreshToken
      tags: [Auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ════════════════════════════════════════
  # Apps (Phase 2)
  # ════════════════════════════════════════
...
  /api/v1/auth/mfa/backup-codes:
    post:
      operationId: regenerateBackupCodes
      tags: [MFA]
      summary: Regenerate backup codes (requires valid TOTP code)
      security: [{BearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MFACodeRequest'
      responses:
        '200':
          description: New backup codes generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MFABackupCodesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ════════════════════════════════════════
  # User Webhooks
  # ════════════════════════════════════════
...
  /api/v1/roles:
    post:
      operationId: createRole
      tags: [Custom Roles]
      summary: Create a custom role (Team+)
      security: [{BearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleInput'
      responses:
        '201':
          description: Role created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomRole'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      operationId: listRoles
      tags: [Custom Roles]
      summary: List custom roles
      security: [{BearerAuth: []}]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomRole'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/roles/permissions:
    get:
      operationId: listPermissions
      tags: [Custom Roles]
      summary: List all available permissions
      security: [{BearerAuth: []}]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Permission'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/roles/{roleId}:
    put:
      operationId: updateRole
      tags: [Custom Roles]
      summary: Update a custom role
      security: [{BearerAuth: []}]
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomRole'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteRole
      tags: [Custom Roles]
      summary: Delete a custom role
      security: [{BearerAuth: []}]
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/roles/{roleId}/assign:
    post:
      operationId: assignRole
      tags: [Custom Roles]
      summary: Assign a role to a team member
      security: [{BearerAuth: []}]
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleInput'
      responses:
        '201':
          description: Role assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleAssignment'
        '401':
          $ref: '#/components/responses/Unauthorized'
...